Refactor app graph
